Split jobs.py
