+++ airflow/task/task_runner/__init__.py
@@ -30,7 +30,7 @@ def get_task_runner(local_task_job):
 
     :param local_task_job: The LocalTaskJob associated with the TaskInstance
     that needs to be executed.
-    :type local_task_job: airflow.jobs.LocalTaskJob
+    :type local_task_job: airflow.jobs.local_task_job.LocalTaskJob
     :return: The task runner to use to run the task.
     :rtype: airflow.task.task_runner.base_task_runner.BaseTaskRunner
     """
