+++ airflow/utils/dag_processing.py
@@ -1216,7 +1216,7 @@ class DagFileProcessorManager(LoggingMixin):
         if (now - self._last_zombie_query_time).total_seconds() \
                 > self._zombie_query_interval:
             # to avoid circular imports
-            from airflow.jobs import LocalTaskJob as LJ
+            from airflow.jobs.local_task_job import LocalTaskJob as LJ
             self.log.info("Finding 'running' jobs without a recent heartbeat")
             TI = airflow.models.TaskInstance
             limit_dttm = timezone.utcnow() - timedelta(
