+++ scripts/perf/scheduler_ops_metrics.py
@@ -22,7 +22,7 @@ import pandas as pd
 import sys
 
 from airflow import configuration, settings
-from airflow.jobs import SchedulerJob
+from airflow.jobs.scheduler_job import SchedulerJob
 from airflow.models import DagBag, DagModel, DagRun, TaskInstance
 from airflow.utils import timezone
 from airflow.utils.state import State
