+++ tests/task/task_runner/test_standard_task_runner.py
@@ -23,7 +23,7 @@ import time
 import unittest
 
 from airflow import models, settings
-from airflow.jobs import LocalTaskJob
+from airflow.jobs.local_task_job import LocalTaskJob
 from airflow.models import TaskInstance as TI
 from airflow.task.task_runner import StandardTaskRunner
 from airflow.utils import timezone
