+++ tests/test_jobs.py
@@ -44,7 +44,10 @@ from airflow import configuration
 from airflow.bin import cli
 import airflow.example_dags
 from airflow.executors import BaseExecutor, SequentialExecutor
-from airflow.jobs import BaseJob, BackfillJob, SchedulerJob, LocalTaskJob
+from airflow.jobs import BaseJob
+from airflow.jobs.local_task_job import LocalTaskJob
+from airflow.jobs.backfill_job import BackfillJob
+from airflow.jobs.scheduler_job import SchedulerJob
 from airflow.models import DAG, DagModel, DagBag, DagRun, Pool, TaskInstance as TI
 from airflow.operators.bash_operator import BashOperator
 from airflow.operators.dummy_operator import DummyOperator
